+++ src/square/mod.rs
@@ -1,11 +1,11 @@
+use crate::fixedvec2d::FixedVec2D;
 use fixedbitset::FixedBitSet;
 use itertools::*;
 use petgraph::{
     data::{DataMap, DataMapMut},
     graph::IndexType,
     visit::{
-        Data, EdgeRef, GraphBase, GraphProp, IntoEdgeReferences, IntoEdges, IntoNeighbors,
-        IntoNodeIdentifiers, IntoNodeReferences, NodeCompactIndexable, NodeCount, NodeIndexable,
+        Data, GraphBase, GraphProp, IntoNodeIdentifiers, IntoNodeReferences, NodeIndexable,
         VisitMap, Visitable,
     },
     Undirected,
@@ -14,69 +14,17 @@ use std::{
     iter::FusedIterator, marker::PhantomData, num::NonZeroUsize, ops::Range, slice::Iter, usize,
 };
 
-use crate::fixedvec2d::FixedVec2D;
+mod edges;
+pub use edges::*;
+mod neighbors;
+pub use neighbors::*;
+mod nodes;
+pub use nodes::*;
+mod index;
+pub use index::*;
 
-/// Axis of the Square grid.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum Axis {
-    Horizontal,
-    Vertical,
-}
-
-impl Axis {
-    /// Check whether axis is horizontal.
-    pub fn is_horizontal(&self) -> bool {
-        *self == Axis::Horizontal
-    }
-    /// Check whether axis is vertical.
-    pub fn is_vertical(&self) -> bool {
-        *self == Axis::Vertical
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum SquareDirection {
-    Foward(Axis),
-    Backward(Axis),
-}
-
-impl SquareDirection {
-    /// Foward Horizontal
-    pub const fn up() -> Self {
-        Self::Foward(Axis::Vertical)
-    }
-    /// Backward Horizontal
-    pub const fn down() -> Self {
-        Self::Backward(Axis::Vertical)
-    }
-    /// Backward Vertical
-    pub const fn left() -> Self {
-        Self::Backward(Axis::Horizontal)
-    }
-    /// Foward Vertical
-    pub const fn right() -> Self {
-        Self::Foward(Axis::Horizontal)
-    }
-}
-
-impl From<(Axis, bool)> for SquareDirection {
-    fn from((axis, dir): (Axis, bool)) -> Self {
-        if dir {
-            SquareDirection::Foward(axis)
-        } else {
-            SquareDirection::Backward(axis)
-        }
-    }
-}
-
-impl From<SquareDirection> for (Axis, bool) {
-    fn from(dir: SquareDirection) -> Self {
-        match dir {
-            SquareDirection::Backward(x) => (x, false),
-            SquareDirection::Foward(x) => (x, true),
-        }
-    }
-}
+#[cfg(test)]
+mod tests;
 
 /// Undirected Square Grid Graph.
 /// ```text
@@ -334,115 +282,6 @@ where
     }
 }
 
-/// Node index for [`SquareGraph`]
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct NodeIndex<Ix: IndexType> {
-    pub horizontal: Ix,
-    pub vertical: Ix,
-}
-
-impl<Ix: IndexType> NodeIndex<Ix> {
-    /// Create a Index from horizontal and vertical.
-    pub fn new(horizontal: Ix, vertical: Ix) -> Self {
-        Self {
-            horizontal,
-            vertical,
-        }
-    }
-
-    /// Returns the manhattan distance
-    pub fn distance<T: Into<(usize, usize)>>(&self, target: T) -> usize {
-        let target: (usize, usize) = target.into();
-        (self.horizontal.index() as isize - target.0 as isize).abs() as usize
-            + (self.vertical.index() as isize - target.1 as isize).abs() as usize
-    }
-
-    /// Get the edge from this node. This does not check whether the node is valid in graph.
-    pub unsafe fn get_edge_id_unchecked(&self, dir: SquareDirection) -> EdgeIndex<Ix> {
-        match dir {
-            SquareDirection::Foward(x) => (*self, x),
-            SquareDirection::Backward(a @ Axis::Vertical) => (
-                Self::new(self.horizontal, Ix::new(self.vertical.index() - 1)),
-                a,
-            ),
-            SquareDirection::Backward(a @ Axis::Horizontal) => (
-                Self::new(Ix::new(self.horizontal.index() - 1), self.vertical),
-                a,
-            ),
-        }
-        .into()
-    }
-
-    #[inline]
-    pub fn up(self) -> Self {
-        Self {
-            vertical: Ix::new(self.vertical.index() + 1),
-            ..self
-        }
-    }
-
-    #[inline]
-    pub fn down(self) -> Self {
-        Self {
-            vertical: Ix::new(self.vertical.index() - 1),
-            ..self
-        }
-    }
-
-    #[inline]
-    pub fn right(self) -> Self {
-        Self {
-            horizontal: Ix::new(self.horizontal.index() + 1),
-            ..self
-        }
-    }
-
-    #[inline]
-    pub fn left(self) -> Self {
-        Self {
-            horizontal: Ix::new(self.horizontal.index() - 1),
-            ..self
-        }
-    }
-}
-
-impl<Ix: IndexType> PartialEq<(usize, usize)> for NodeIndex<Ix> {
-    fn eq(&self, value: &(usize, usize)) -> bool {
-        &(self.horizontal.index(), self.vertical.index()) == value
-    }
-}
-
-impl<Ix: IndexType> From<(usize, usize)> for NodeIndex<Ix> {
-    fn from(value: (usize, usize)) -> Self {
-        NodeIndex::new(Ix::new(value.0), Ix::new(value.1))
-    }
-}
-
-impl<Ix: IndexType> From<NodeIndex<Ix>> for (usize, usize) {
-    fn from(value: NodeIndex<Ix>) -> Self {
-        (value.horizontal.index(), value.vertical.index())
-    }
-}
-
-/// Edge Index of [`SquareGraph`]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct EdgeIndex<Ix: IndexType> {
-    pub node: NodeIndex<Ix>,
-    pub axis: Axis,
-}
-
-impl<Ix: IndexType> From<(NodeIndex<Ix>, Axis)> for EdgeIndex<Ix> {
-    fn from((n, a): (NodeIndex<Ix>, Axis)) -> Self {
-        Self { node: n, axis: a }
-    }
-}
-
-impl<Ix: IndexType> From<(NodeIndex<Ix>, SquareDirection)> for EdgeIndex<Ix> {
-    fn from((n, a): (NodeIndex<Ix>, SquareDirection)) -> Self {
-        unsafe { n.get_edge_id_unchecked(a) }
-    }
-}
-
 impl<N, E, Ix> GraphBase for SquareGraph<N, E, Ix>
 where
     Ix: IndexType,
@@ -510,461 +349,6 @@ where
     type EdgeType = Undirected;
 }
 
-impl<'a, N, E, Ix> IntoEdgeReferences for &'a SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-    E: Copy,
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type EdgeRef = EdgeReference<'a, E, Ix>;
-    type EdgeReferences = EdgeReferences<'a, E, Ix>;
-
-    fn edge_references(self) -> Self::EdgeReferences {
-        EdgeReferences::new(&self)
-    }
-}
-
-/// Reference of Edge data (EdgeIndex, EdgeWeight, direction) in [`SquareGraph`].
-#[derive(Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct EdgeReference<'a, E, Ix: IndexType> {
-    edge_id: EdgeIndex<Ix>,
-    edge_weight: &'a E,
-    direction: bool,
-}
-
-impl<'a, E, Ix: IndexType> Clone for EdgeReference<'a, E, Ix> {
-    fn clone(&self) -> Self {
-        Self {
-            edge_id: self.edge_id,
-            edge_weight: self.edge_weight,
-            direction: self.direction,
-        }
-    }
-}
-
-impl<'a, E, Ix: IndexType> Copy for EdgeReference<'a, E, Ix> {}
-
-impl<'a, E, Ix: IndexType> EdgeReference<'a, E, Ix> {
-    #[inline]
-    fn get_node(&self, is_source: bool) -> NodeIndex<Ix> {
-        if is_source {
-            (self.edge_id).node
-        } else {
-            match (self.edge_id).axis {
-                Axis::Horizontal => (self.edge_id).node.right(),
-                Axis::Vertical => (self.edge_id).node.up(),
-            }
-        }
-    }
-}
-
-impl<'a, E: Copy, Ix: IndexType> EdgeRef for EdgeReference<'a, E, Ix> {
-    type NodeId = NodeIndex<Ix>;
-    type EdgeId = EdgeIndex<Ix>;
-    type Weight = E;
-
-    #[inline]
-    fn source(&self) -> Self::NodeId {
-        self.get_node(self.direction)
-    }
-
-    #[inline]
-    fn target(&self) -> Self::NodeId {
-        self.get_node(!self.direction)
-    }
-
-    fn weight(&self) -> &Self::Weight {
-        self.edge_weight
-    }
-
-    fn id(&self) -> Self::EdgeId {
-        self.edge_id
-    }
-}
-
-/// Iterator for all edges of [`SquareGraph`].
-#[derive(Clone, Debug)]
-pub struct EdgeReferences<'a, E, Ix: IndexType> {
-    horizontal: &'a FixedVec2D<E>,
-    vertical: &'a FixedVec2D<E>,
-    nodes: NodeIndices<Ix>,
-    prv: Option<EdgeIndex<Ix>>,
-}
-
-impl<'a, E, Ix: IndexType> EdgeReferences<'a, E, Ix> {
-    fn new<N>(graph: &'a SquareGraph<N, E, Ix>) -> Self {
-        Self {
-            horizontal: &graph.horizontal,
-            vertical: &graph.vertical,
-            nodes: NodeIndices::new(graph.horizontal_node_count(), graph.vertical_node_count()),
-            prv: None,
-        }
-    }
-}
-
-impl<'a, E, Ix> Iterator for EdgeReferences<'a, E, Ix>
-where
-    Ix: IndexType,
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type Item = EdgeReference<'a, E, Ix>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            if let Some(ref mut e) = self.prv {
-                if e.axis == Axis::Horizontal {
-                    let item = self.vertical.ref_2d()[e.node.horizontal.index()]
-                        .get(e.node.vertical.index());
-                    if let Some(item) = item {
-                        e.axis = Axis::Vertical;
-                        return Some(EdgeReference {
-                            edge_id: *e,
-                            edge_weight: item,
-                            direction: true,
-                        });
-                    }
-                }
-            }
-            if let Some(next) = self.nodes.next() {
-                let item = self
-                    .horizontal
-                    .ref_2d()
-                    .get(next.horizontal.index())
-                    .map(|x| x.get(next.vertical.index()))
-                    .flatten();
-                let edge_id = EdgeIndex {
-                    node: next,
-                    axis: Axis::Horizontal,
-                };
-                self.prv = Some(edge_id);
-                if let Some(edge_weight) = item {
-                    return Some(EdgeReference {
-                        edge_id,
-                        edge_weight,
-                        direction: true,
-                    });
-                }
-            } else {
-                return None;
-            }
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let (lo, hi) = self.nodes.size_hint();
-        (
-            lo.saturating_sub(self.horizontal.size() + self.vertical.size()),
-            hi.map(|x| x * 2),
-        )
-    }
-}
-
-pub struct Edges<'a, N, E, Ix: IndexType> {
-    g: &'a SquareGraph<N, E, Ix>,
-    node: NodeIndex<Ix>,
-    state: usize,
-}
-
-impl<'a, N, E, Ix> Iterator for Edges<'a, N, E, Ix>
-where
-    Ix: IndexType,
-{
-    type Item = EdgeReference<'a, E, Ix>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let g = self.g;
-        let n = self.node;
-        loop {
-            'inner: loop {
-                let er = match self.state {
-                    0 => {
-                        if n.horizontal.index() == 0 {
-                            break 'inner;
-                        }
-                        EdgeReference {
-                            edge_id: EdgeIndex {
-                                node: n.left(),
-                                axis: Axis::Horizontal,
-                            },
-                            edge_weight: unsafe {
-                                g.horizontal
-                                    .ref_2d()
-                                    .get_unchecked(n.horizontal.index() - 1)
-                                    .get_unchecked(n.vertical.index())
-                            },
-                            direction: false,
-                        }
-                    }
-                    1 => {
-                        if n.horizontal.index() + 1 >= g.horizontal_node_count() {
-                            break 'inner;
-                        }
-                        EdgeReference {
-                            edge_id: EdgeIndex {
-                                node: n,
-                                axis: Axis::Horizontal,
-                            },
-                            edge_weight: unsafe {
-                                g.horizontal
-                                    .ref_2d()
-                                    .get_unchecked(n.horizontal.index())
-                                    .get_unchecked(n.vertical.index())
-                            },
-                            direction: true,
-                        }
-                    }
-                    2 => {
-                        if n.vertical.index() == 0 {
-                            break 'inner;
-                        }
-                        EdgeReference {
-                            edge_id: EdgeIndex {
-                                node: n.down(),
-                                axis: Axis::Vertical,
-                            },
-                            edge_weight: unsafe {
-                                g.vertical
-                                    .ref_2d()
-                                    .get_unchecked(n.horizontal.index())
-                                    .get_unchecked(n.vertical.index() - 1)
-                            },
-                            direction: false,
-                        }
-                    }
-                    3 => {
-                        if n.vertical.index() + 1 >= g.vertical_node_count() {
-                            break 'inner;
-                        }
-                        EdgeReference {
-                            edge_id: EdgeIndex {
-                                node: n,
-                                axis: Axis::Vertical,
-                            },
-                            edge_weight: unsafe {
-                                g.vertical
-                                    .ref_2d()
-                                    .get_unchecked(n.horizontal.index())
-                                    .get_unchecked(n.vertical.index())
-                            },
-                            direction: true,
-                        }
-                    }
-                    _ => return None,
-                };
-                self.state += 1;
-                return Some(er);
-            }
-            self.state += 1;
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (0, Some(4 - self.state))
-    }
-}
-
-impl<'a, N, E, Ix> FusedIterator for Edges<'a, N, E, Ix> where Ix: IndexType {}
-
-impl<'a, N, E, Ix> IntoEdges for &'a SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-    E: Copy,
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type Edges = Edges<'a, N, E, Ix>;
-
-    fn edges(self, a: Self::NodeId) -> Self::Edges {
-        Edges {
-            g: &self,
-            node: a,
-            state: 0,
-        }
-    }
-}
-
-pub struct Neighbors<Ix: IndexType> {
-    node: NodeIndex<Ix>,
-    state: usize,
-    h: usize,
-    v: usize,
-}
-
-impl<Ix> Iterator for Neighbors<Ix>
-where
-    Ix: IndexType,
-{
-    type Item = NodeIndex<Ix>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            let n = self.node;
-            let x = match self.state {
-                0 if n.horizontal.index() != 0 => n.left(),
-                1 if n.horizontal.index() + 1 < self.h => n.right(),
-                2 if n.vertical.index() != 0 => n.down(),
-                3 if n.vertical.index() + 1 < self.v => n.up(),
-                4..=usize::MAX => return None,
-                _ => {
-                    self.state += 1;
-                    continue;
-                }
-            };
-            self.state += 1;
-            return Some(x);
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (0, Some(4 - self.state))
-    }
-}
-
-impl<Ix> FusedIterator for Neighbors<Ix> where Ix: IndexType {}
-
-impl<'a, N, E, Ix> IntoNeighbors for &'a SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-{
-    type Neighbors = Neighbors<Ix>;
-
-    fn neighbors(self: Self, a: Self::NodeId) -> Self::Neighbors {
-        Neighbors {
-            node: a,
-            state: 0,
-            h: self.horizontal_node_count(),
-            v: self.vertical_node_count(),
-        }
-    }
-}
-
-impl<'a, N, E, Ix> IntoNodeIdentifiers for &'a SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type NodeIdentifiers = NodeIndices<Ix>;
-
-    fn node_identifiers(self) -> Self::NodeIdentifiers {
-        NodeIndices::new(self.horizontal_node_count(), self.vertical_node_count())
-    }
-}
-
-/// Iterate all index of [`SquareGraph`].
-#[derive(Clone, Debug)]
-pub struct NodeIndices<Ix> {
-    p: itertools::Product<Range<usize>, Range<usize>>,
-    pd: PhantomData<Ix>,
-}
-
-impl<Ix> NodeIndices<Ix> {
-    fn new(h: usize, v: usize) -> Self {
-        Self {
-            p: (0..h).cartesian_product(0..v),
-            pd: PhantomData,
-        }
-    }
-}
-
-impl<Ix: IndexType> Iterator for NodeIndices<Ix>
-where
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type Item = NodeIndex<Ix>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.p.next().map(|x| (x.0, x.1).into())
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.p.size_hint()
-    }
-
-    fn fold<B, F>(self, init: B, mut f: F) -> B
-    where
-        Self: Sized,
-        F: FnMut(B, Self::Item) -> B,
-    {
-        self.p.fold(init, |x, item| (&mut f)(x, item.into()))
-    }
-}
-
-impl<Ix: IndexType> FusedIterator for NodeIndices<Ix> where Range<Ix>: Iterator<Item = Ix> {}
-
-impl<'a, N: Clone, E, Ix> IntoNodeReferences for &'a SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type NodeRef = (NodeIndex<Ix>, &'a N);
-    type NodeReferences = NodeReferences<'a, N, Ix>;
-
-    fn node_references(self) -> Self::NodeReferences {
-        NodeReferences {
-            indices: self.node_identifiers(),
-            nodes: self.nodes.ref_1d().iter(),
-        }
-    }
-}
-
-/// Iterate all nodes of [`SquareGraph`].
-pub struct NodeReferences<'a, N, Ix> {
-    indices: NodeIndices<Ix>,
-    nodes: Iter<'a, N>,
-}
-
-impl<'a, N, Ix> Iterator for NodeReferences<'a, N, Ix>
-where
-    Ix: IndexType,
-    Range<Ix>: Iterator<Item = Ix>,
-{
-    type Item = (NodeIndex<Ix>, &'a N);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let n = self.nodes.next()?;
-        Some((self.indices.next()?, n))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.nodes.size_hint()
-    }
-
-    fn count(self) -> usize
-    where
-        Self: Sized,
-    {
-        self.nodes.count()
-    }
-}
-
-impl<N, E, Ix> NodeCompactIndexable for SquareGraph<N, E, Ix> where Ix: IndexType {}
-
-impl<N, E, Ix> NodeCount for SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-{
-    fn node_count(self: &Self) -> usize {
-        self.nodes.size()
-    }
-}
-
-impl<N, E, Ix> NodeIndexable for SquareGraph<N, E, Ix>
-where
-    Ix: IndexType,
-{
-    fn node_bound(self: &Self) -> usize {
-        self.nodes.size()
-    }
-
-    fn to_index(self: &Self, a: Self::NodeId) -> usize {
-        a.horizontal.index() * self.vertical_node_count() + a.vertical.index()
-    }
-
-    fn from_index(self: &Self, i: usize) -> Self::NodeId {
-        let h = self.vertical_node_count();
-        (i / h, i % h).into()
-    }
-}
-
 /// [`VisitMap`] of [`SquareGraph`].
 #[derive(Debug, Clone, Hash, PartialEq, Eq)]
 pub struct VisMap {
@@ -1003,183 +387,5 @@ where
 
     fn reset_map(self: &Self, map: &mut Self::Map) {
         map.v.iter_mut().for_each(|x| x.clear())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::array::IntoIter;
-
-    use super::*;
-
-    #[test]
-    fn gen() {
-        let sq = SquareGraph::<_, _, u32>::new_with(
-            4,
-            3,
-            |x, y| x + 2 * y,
-            |x, y, d| (x + 2 * y) as i32 * (if d.is_horizontal() { 1 } else { -1 }),
-        );
-        assert_eq!(sq.horizontal_node_count(), 4);
-        assert_eq!(sq.vertical_node_count(), 3);
-        assert_eq!(sq.node_weight((0, 0).into()), Some(&0));
-        assert_eq!(sq.node_weight((3, 0).into()), Some(&3));
-        assert_eq!(sq.node_weight((4, 0).into()), None);
-        assert_eq!(sq.node_weight((0, 2).into()), Some(&4));
-        assert_eq!(sq.node_weight((0, 3).into()), None);
-        assert_eq!(
-            sq.edge_weight(((0, 0).into(), Axis::Horizontal).into()),
-            Some(&0)
-        );
-        assert_eq!(
-            sq.edge_weight(((0, 2).into(), Axis::Horizontal).into()),
-            Some(&4)
-        );
-        assert_eq!(sq.edge_weight(((0, 2).into(), Axis::Vertical).into()), None);
-        assert_eq!(
-            sq.edge_weight(((3, 0).into(), Axis::Horizontal).into()),
-            None
-        );
-        assert_eq!(
-            sq.edge_weight(((3, 0).into(), Axis::Vertical).into()),
-            Some(&-3)
-        );
-    }
-
-    #[test]
-    fn node_identifiers() {
-        let sq = SquareGraph::<_, _, u32>::new_with(
-            3,
-            5,
-            |x, y| x + 2 * y,
-            |x, y, d| (x + 2 * y) as i32 * (if d.is_horizontal() { 1 } else { -1 }),
-        );
-        for (i, x) in sq.node_identifiers().enumerate() {
-            let x = x;
-            let x2 = sq.to_index(x);
-            assert_eq!(x2, i);
-            let x3 = sq.from_index(x2);
-            assert_eq!(x, x3);
-        }
-    }
-
-    #[test]
-    fn neighbors() {
-        let sq = SquareGraph::<_, _, u32>::new_with(
-            3,
-            5,
-            |x, y| x + 2 * y,
-            |x, y, d| (x + 2 * y) as i32 * (if d.is_horizontal() { 1 } else { -1 }),
-        );
-
-        let v00 = sq.neighbors((0, 0).into());
-        debug_assert!(v00.eq(IntoIter::new([(1, 0), (0, 1)])));
-
-        let v04 = sq.neighbors((0, 4).into());
-        debug_assert!(v04.eq(IntoIter::new([(1, 4), (0, 3)])));
-
-        let v20 = sq.neighbors((2, 0).into());
-        debug_assert!(v20.eq(IntoIter::new([(1, 0), (2, 1)])));
-
-        let v24 = sq.neighbors((2, 4).into());
-        debug_assert!(v24.eq(IntoIter::new([(1, 4), (2, 3)])));
-
-        let v12 = sq.neighbors((1, 2).into());
-        debug_assert!(v12.eq(IntoIter::new([(0, 2), (2, 2), (1, 1), (1, 3)])));
-    }
-
-    #[test]
-    fn edges() {
-        let sq = SquareGraph::<_, _, u32>::new_with(
-            3,
-            5,
-            |x, y| x + 2 * y,
-            |x, y, d| (x + 2 * y) as i32 * (if d.is_horizontal() { 1 } else { -1 }),
-        );
-
-        debug_assert!(sq
-            .edges((0, 0).into())
-            .map(|e| e.target())
-            .eq(IntoIter::new([(1, 0), (0, 1)])));
-
-        debug_assert!(sq.edges((0, 0).into()).map(|e| e.edge_weight).eq(&[0, 0]));
-        debug_assert!(sq
-            .edges((1, 1).into())
-            .map(|e| e.edge_weight)
-            .eq(&[2, 3, -1, -3]));
-
-        debug_assert!(sq
-            .edges((1, 2).into())
-            .map(|e| e.target())
-            .eq(IntoIter::new([(0, 2), (2, 2), (1, 1), (1, 3)])));
-    }
-
-    #[test]
-    fn edge_references() {
-        let sq = SquareGraph::<_, _, u32>::new_with(
-            3,
-            5,
-            |x, y| x + 2 * y,
-            |x, y, d| (x + 2 * y) as i32 * (if d.is_horizontal() { 1 } else { -1 }),
-        );
-
-        let mut i = 0;
-        let mut x = -1;
-        for e in sq
-            .edge_references()
-            .filter(|x| x.id().axis == Axis::Horizontal)
-        {
-            let y = sq.to_index(e.edge_id.node) as i32;
-            assert!(x < y);
-            x = y;
-            i += 1;
-        }
-        assert_eq!(i, 10);
-        x = -1;
-        i = 0;
-        for e in sq
-            .edge_references()
-            .filter(|x| x.id().axis == Axis::Vertical)
-        {
-            let y = sq.to_index(e.edge_id.node) as i32;
-            assert!(x < y);
-            x = y;
-            i += 1;
-        }
-        assert_eq!(i, 12);
-    }
-
-    #[test]
-    fn astar() {
-        let sq = SquareGraph::<_, _, u32>::new_with(
-            3,
-            4,
-            |_, _| (),
-            |x, y, d| (x + 2 * y) as i32 * (if d.is_horizontal() { 1 } else { 3 }),
-        );
-
-        let x = petgraph::algo::astar(
-            &sq,
-            (0, 0).into(),
-            |x| x == (2, 1),
-            |e| *e.weight(),
-            |x| x.distance((2, 1)) as i32,
-        );
-        assert!(x.is_some());
-        let (d, p) = x.unwrap();
-        assert_eq!(d, 5);
-        assert_eq!(p, [(0, 0), (0, 1), (1, 1), (2, 1)]);
-
-        let x = petgraph::algo::astar(
-            &sq,
-            (2, 1).into(),
-            |x| x == (0, 0),
-            |e| *e.weight(),
-            |x| x.distance((0, 0)) as i32,
-        );
-        assert!(x.is_some());
-        let (d, p) = x.unwrap();
-        assert_eq!(d, 5);
-        assert_eq!(p, [(2, 1), (1, 1), (0, 1), (0, 0)])
     }
 }
